Add target_features to TransFnAttrs
Part of #47320
